pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eo-pipeline"
version = "1.0.0"
description = "Earth Observation AI Pipeline for satellite imagery analysis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "EO Pipeline Team", email = "team@example.com"}
]
keywords = [
    "earth-observation",
    "satellite-imagery",
    "deep-learning",
    "pytorch",
    "remote-sensing",
    "land-use-classification",
    "mlops"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "rasterio>=1.3.0",
    "geopandas>=0.14.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
mlops = [
    "mlflow>=2.9.0",
    "dvc>=3.30.0",
    "dvc-s3>=3.0.0",
]
web = [
    "gradio>=4.10.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]
xai = [
    "shap>=0.43.0",
    "lime>=0.2.0",
    "captum>=0.6.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]
all = [
    "eo-pipeline[mlops,web,xai,dev]",
]

[project.scripts]
eo-pipeline = "eo_pipeline.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/octo"
Documentation = "https://github.com/yourusername/octo#readme"
Repository = "https://github.com/yourusername/octo.git"
Issues = "https://github.com/yourusername/octo/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["eo_pipeline*", "mlops*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | __pycache__
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["eo_pipeline", "mlops"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "build/",
    "dist/",
    "venv/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["eo_pipeline"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
